math: Fix fp_traits for IBM long double
This patch fixes the fp_traits_non_native template specialization for
IBM long double.  Current interface does not work correctly for the type
due to the fact that for some operations (sign manipulation, for instance), both
double should be manipulated.  For other operations that only require
the 32 most significand bits (isinf/isnan for instance), the double
specialization is suffient.

This fixes the issues for geometry test in PPC64LE environment with no
regressions on PPC64BE.